Finished my cool shopping list!
All finished with testing and everything. The new feature of suggestions has been fully completed and I even added the feature of not being able to add an item to the list multiple times. Super stoked about this. Now I want to learn how to turn this into a mobile app so I can actually use it!!